build_docker_rotation/Rotation/requirements.txt
torch==1.6.0
absl-py==0.10.0
aiofiles==0.5.0
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
blis==0.4.1
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
commonmark==0.9.1
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
decorator==4.4.2
editdistance==0.3.1
efficientnet==1.0.0
efficientnet-pytorch==0.7.0
essential-generators==0.9.2
fastapi==0.44.1
Flask==1.1.2
fonttools==4.16.1
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.2
grpcio==1.32.0
h11==0.9.0
h5py==2.10.0
httptools==0.1.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==2.0.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.2.0
lmdb==1.0.0
Markdown==3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
murmurhash==1.0.2
networkx==2.5
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
objgraph==3.4.1
opencv-python==4.2.0.32
opt-einsum==3.3.0
paddleocr==1.0.0
paddlepaddle==1.8.5
parso==0.7.1
pathlib==1.0.1
peewee==3.13.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
portalocker==2.0.0
preshed==3.0.2
prettytable==1.0.0
prompt-toolkit==3.0.7
protobuf==3.1.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.2.0
pycocotools==2.0.2
pydantic==1.6.1
pydot==1.4.1
Pygments==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
rarfile==4.0
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
rich==8.0.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.2
Shapely==1.7.1
six==1.15.0
spacy==2.3.2
srsly==1.0.2
starlette==0.12.9
tabulate==0.8.7
termcolor==1.1.0
toolz==0.11.1
torchvision
tqdm==4.50.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
validators==0.18.1
wasabi==0.8.0
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yacs==0.1.8
zipp==3.3.0
